Moving sliceutil inside slice
