+++ pkg/autocomplete/generator.go
@@ -26,7 +26,7 @@ import (
 	"github.com/ZupIT/ritchie-cli/pkg/api"
 	"github.com/ZupIT/ritchie-cli/pkg/formula"
 	"github.com/ZupIT/ritchie-cli/pkg/prompt"
-	"github.com/ZupIT/ritchie-cli/pkg/sliceutil"
+	"github.com/ZupIT/ritchie-cli/pkg/slice/sliceutil"
 )
 
 const (
